Server/requirements.txt
google.generativeai
pinecone-client
mixedbread-ai
pylatexenc
langchain
pypdf
html2text
dotenv
tiktoken
pandas
numpy